TST: Finish removal of groupby/test_function.py
