+++ pandas/tests/groupby/test_numeric_only.py
@@ -205,39 +205,6 @@ class TestNumericOnly:
             tm.assert_index_equal(result.columns, expected_columns)
 
 
-@pytest.mark.parametrize(
-    "i",
-    [
-        (
-            Timestamp("2011-01-15 12:50:28.502376"),
-            Timestamp("2011-01-20 12:50:28.593448"),
-        ),
-        (24650000000000001, 24650000000000002),
-    ],
-)
-def test_groupby_non_arithmetic_agg_int_like_precision(i):
-    # see gh-6620, gh-9311
-    df = DataFrame([{"a": 1, "b": i[0]}, {"a": 1, "b": i[1]}])
-
-    grp_exp = {
-        "first": {"expected": i[0]},
-        "last": {"expected": i[1]},
-        "min": {"expected": i[0]},
-        "max": {"expected": i[1]},
-        "nth": {"expected": i[1], "args": [1]},
-        "count": {"expected": 2},
-    }
-
-    for method, data in grp_exp.items():
-        if "args" not in data:
-            data["args"] = []
-
-        grouped = df.groupby("a")
-        res = getattr(grouped, method)(*data["args"])
-
-        assert res.iloc[0].b == data["expected"]
-
-
 @pytest.mark.parametrize("numeric_only", [True, False, None])
 def test_axis1_numeric_only(request, groupby_func, numeric_only):
     if groupby_func in ("idxmax", "idxmin"):
@@ -543,43 +510,3 @@ def test_deprecate_numeric_only_series(dtype, groupby_func, request):
         result = method(*args, numeric_only=True)
         expected = method(*args, numeric_only=False)
         tm.assert_series_equal(result, expected)
-
-
-def test_multiindex_group_all_columns_when_empty(groupby_func):
-    # GH 32464
-    df = DataFrame({"a": [], "b": [], "c": []}).set_index(["a", "b", "c"])
-    gb = df.groupby(["a", "b", "c"], group_keys=False)
-    method = getattr(gb, groupby_func)
-    args = get_groupby_method_args(groupby_func, df)
-
-    warn = FutureWarning if groupby_func == "fillna" else None
-    warn_msg = "DataFrameGroupBy.fillna is deprecated"
-    with tm.assert_produces_warning(warn, match=warn_msg):
-        result = method(*args).index
-    expected = df.index
-    tm.assert_index_equal(result, expected)
-
-
-def test_duplicate_columns(request, groupby_func, as_index):
-    # GH#50806
-    if groupby_func == "corrwith":
-        msg = "GH#50845 - corrwith fails when there are duplicate columns"
-        request.applymarker(pytest.mark.xfail(reason=msg))
-    df = DataFrame([[1, 3, 6], [1, 4, 7], [2, 5, 8]], columns=list("abb"))
-    args = get_groupby_method_args(groupby_func, df)
-    gb = df.groupby("a", as_index=as_index)
-    warn = FutureWarning if groupby_func == "fillna" else None
-    warn_msg = "DataFrameGroupBy.fillna is deprecated"
-    with tm.assert_produces_warning(warn, match=warn_msg):
-        result = getattr(gb, groupby_func)(*args)
-
-    expected_df = df.set_axis(["a", "b", "c"], axis=1)
-    expected_args = get_groupby_method_args(groupby_func, expected_df)
-    expected_gb = expected_df.groupby("a", as_index=as_index)
-    warn = FutureWarning if groupby_func == "fillna" else None
-    warn_msg = "DataFrameGroupBy.fillna is deprecated"
-    with tm.assert_produces_warning(warn, match=warn_msg):
-        expected = getattr(expected_gb, groupby_func)(*expected_args)
-    if groupby_func not in ("size", "ngroup", "cumcount"):
-        expected = expected.rename(columns={"c": "b"})
-    tm.assert_equal(result, expected)
